FastAPI/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
numpy
hyperon
